Adds version-agnostic layoutMargins property to UIView; adds bottom bleed anchor and 'bleed area' to UIViewController; moves BoundingAnchor types into a group for better organization.
